fix(libraries, thermocycler): switch ADS1X15 adc library, fix TC max() bug (#67)
the new library fixes previous lib's incorrect conversion delay. Also fixed a bug in TC thermistors
`_read_adc` where max() function caused the thermistors to be read twice, hence hiding the bug from the previous library.

+++ modules/thermo-cycler/thermo-cycler-arduino/thermistorsadc.h
@@ -77,8 +77,8 @@ class ThermistorsADC{
             Adafruit_ADS1115 *adc_a;
             Adafruit_ADS1115 *adc_b;
 
-            int _read_adc(int index);
-            float _adc_to_celsius(int _adc);
+            uint16_t _read_adc(int index);
+            float _adc_to_celsius(uint16_t _adc);
 
             double probe_temps[TOTAL_THERMISTORS] = {0, };
             double sum_probe_temps[TOTAL_THERMISTORS] = {0, };
